removed uneccessary files
