--- controllers/enum.js
@@ -1,9 +1,0 @@
-const VERIFICATION_TYPE = {
-    MOBILE: "MOBILE",
-    EMAIL: "EMAIL",
-    DOCUMENT: "DOCUMENT",
-}
-
-module.exports = {
-    VERIFICATION_TYPE
-}
