+++ lib/pageRoutes.js
@@ -1,5 +1,3 @@
-const { updatePassword } = require("../controllers/userCtrl");
-
 const pageRoutes = {
     auth: {
         confirmEmail: "/confirm-email",
