+++ middlewares/authMiddleware.js
@@ -1,7 +1,7 @@
-const User = require("../models/userModel");
+const { User } = require("../modules/users/models");
 const asyncHandler = require("express-async-handler");
 const jwt = require("jsonwebtoken");
-const { ROLES } = require("../models/enums");
+const { ROLES } = require("../modules/users/enums");
 
 const authMiddleware = asyncHandler(async (req, res, next) => {
   let token;
