+++ modules/auth/contollers.js
@@ -4,7 +4,7 @@ const jwt = require('jsonwebtoken');
 const { generateToken, verificationToken } = require('../../config/jwtToken');
 const { generateRefreshToken } = require('../../config/refreshToken');
 const { sendEmail, loadTemplate } = require('../../helpers/emailHelper');
-const { ROLES } = require('../../models/enums');
+const { ROLES } = require('../users/enums');
 const { FORGOT_PASSWORD, forgotPasswordValues } = require('../../helpers/mail_templates/forgotPassword');
 const { emailVerificationValues } = require('../../helpers/mail_templates/emailVerification');
 const { pageRoutes } = require('../../lib/pageRoutes');
